add session and fetch edge case handling

+++ app/api/follow/route.ts
@@ -3,10 +3,14 @@ import { prisma } from "@/lib/prisma";
 import { getServerSession } from "next-auth";
 import { NextRequest, NextResponse } from "next/server";
 
-export async function POST(req: Request) {
+export async function POST(req: NextRequest) {
   const session = await getServerSession(authOptions)
   const { targetUserId } = await req.json()
-  const currentUserId = session?.user?.id!
+  const currentUserId = session?.user?.id
+
+  if (!currentUserId) {
+    return NextResponse.redirect("/api/auth/signin")
+  }
 
   const record = await prisma.follows.create({
     data: {
@@ -20,8 +24,12 @@ export async function POST(req: Request) {
 
 export async function DELETE(req: NextRequest) {
   const session = await getServerSession(authOptions)
-  const currentUserId = session?.user?.id!
+  const currentUserId = session?.user?.id
   const targetUserId = req.nextUrl.searchParams.get("targetUserId")
+
+  if (!currentUserId) {
+    return NextResponse.redirect("/api/auth/signin")
+  }
 
   const record = await prisma.follows.delete({
     where: {
